@@ -20,102 +23,206 @@
   "license": "Apache-2.0",
   "author": "Rashid Khan <rashid.khan@elastic.co>",
   "contributors": [
-    "Spencer Alger <spencer.alger@elastic.co>",
     "Chris Cowan <chris.cowan@elastic.co>",
+    "Court Ewing <court@elastic.co>",
+    "Jim Unger <jim.unger@elastic.co>",
     "Joe Fleming <joe.fleming@elastic.co>",
-    "Lukas Olson <lukas.olson@elastic.co>",
+    "Jon Budzenski <jonathan.budzenski@elastic.co>",
     "Juan Thomassie <juan.thomassie@elastic.co>",
+    "Khalah Jones-Golden <khalah.jones@elastic.co>",
+    "Lee Drengenberg <lee.drengenberg@elastic.co>",
+    "Lukas Olson <lukas.olson@elastic.co>",
+    "Matt Bargar <matt.bargar@elastic.co>",
+    "Nicolás Bevacqua <nico@elastic.co>",
     "Shelby Sturgis <shelby@elastic.co>",
-    "Khalah Jones-Golden <khalah.jones@elastic.co>"
+    "Spencer Alger <spencer.alger@elastic.co>",
+    "Tim Sullivan <tim@elastic.co>"
   ],
   "scripts": {
     "test": "grunt test",
-    "start": "node ./src/server/bin/kibana.js",
-    "postinstall": "bower install && grunt licenses --check-validity",
-    "precommit": "grunt lintStagedFiles"
+    "test:dev": "grunt test:dev",
+    "test:quick": "grunt test:quick",
+    "test:browser": "grunt test:browser",
+    "test:ui": "grunt test:ui",
+    "test:ui:server": "grunt test:ui:server",
+    "test:ui:runner": "grunt test:ui:runner",
+    "test:server": "grunt test:server",
+    "test:coverage": "grunt test:coverage",
+    "test:visualRegression": "grunt test:visualRegression",
+    "checkLicenses": "grunt licenses",
+    "build": "grunt build",
+    "release": "grunt release",
+    "start": "sh ./bin/kibana --dev",
+    "precommit": "grunt precommit",
+    "karma": "karma start",
+    "elasticsearch": "grunt esvm:dev:keepalive",
+    "elasticsearchWithPlugins": "grunt esvm:withPlugins:keepalive",
+    "lint": "grunt eslint:source",
+    "lintroller": "grunt eslint:fixSource",
+    "makelogs": "makelogs",
+    "mocha": "mocha",
+    "mocha:debug": "mocha --debug-brk",
+    "sterilize": "grunt sterilize"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/elastic/kibana.git"
   },
   "dependencies": {
-    "ansicolors": "^0.3.2",
-    "bluebird": "^2.9.27",
-    "body-parser": "^1.10.1",
-    "bunyan": "^1.2.3",
-    "commander": "^2.6.0",
-    "compression": "^1.3.0",
-    "cookie-parser": "^1.3.3",
-    "debug": "^2.1.1",
-    "elasticsearch": "^5.0.0",
-    "express": "^4.10.6",
-    "glob": "^4.3.2",
-    "good": "^5.1.2",
-    "good-console": "^4.1.0",
-    "good-file": "^4.0.2",
-    "good-reporter": "^3.1.0",
-    "hapi": "^8.6.1",
-    "http-auth": "^2.2.5",
-    "joi": "^6.4.3",
-    "js-yaml": "^3.2.5",
-    "lodash": "^3.9.3",
-    "json-stringify-safe": "^5.0.1",
-    "moment": "^2.10.3",
-    "numeral": "^1.5.3",
-    "request": "^2.40.0",
-    "requirefrom": "^0.2.0",
-    "semver": "^4.3.6",
-    "serve-favicon": "^2.2.0",
-    "through": "^2.3.6"
+    "@bigfunger/decompress-zip": "0.2.0-stripfix3",
+    "@bigfunger/jsondiffpatch": "0.1.38-webpack",
+    "@elastic/datemath": "2.3.0",
+    "@elastic/kibana-ui-framework": "0.0.11",
+    "@spalger/filesaver": "1.1.2",
+    "@spalger/leaflet-draw": "0.2.3",
+    "@spalger/leaflet-heat": "0.1.3",
+    "@spalger/numeral": "^2.0.0",
+    "@spalger/test-subj-selector": "0.2.1",
+    "@spalger/ui-ace": "0.2.3",
+    "JSONStream": "1.1.1",
+    "angular": "1.4.7",
+    "angular-bootstrap-colorpicker": "3.0.19",
+    "angular-elastic": "2.5.0",
+    "angular-route": "1.4.7",
+    "angular-sanitize": "1.5.7",
+    "ansicolors": "0.3.2",
+    "autoprefixer": "6.3.7",
+    "autoprefixer-loader": "2.0.0",
+    "babel": "5.8.38",
+    "babel-core": "5.8.38",
+    "babel-loader": "5.3.2",
+    "babel-runtime": "5.8.38",
+    "bluebird": "2.9.34",
+    "body-parser": "1.12.0",
+    "boom": "2.8.0",
+    "brace": "0.5.1",
+    "bunyan": "1.7.1",
+    "check-hash": "1.0.1",
+    "commander": "2.8.1",
+    "css-loader": "0.17.0",
+    "d3": "3.5.6",
+    "d3-cloud": "1.2.1",
+    "dragula": "3.7.0",
+    "elasticsearch": "12.0.0-rc5",
+    "elasticsearch-browser": "12.0.0-rc5",
+    "encode-uri-query": "1.0.0",
+    "even-better": "7.0.2",
+    "expiry-js": "0.1.7",
+    "exports-loader": "0.6.2",
+    "expose-loader": "0.7.0",
+    "extract-text-webpack-plugin": "0.8.2",
+    "file-loader": "0.8.4",
+    "font-awesome": "4.4.0",
+    "glob": "5.0.13",
+    "glob-all": "3.0.1",
+    "good-squeeze": "2.1.0",
+    "gridster": "0.5.6",
+    "h2o2": "5.1.1",
+    "hapi": "14.2.0",
+    "httpolyglot": "0.1.1",
+    "imports-loader": "0.6.4",
+    "inert": "4.0.2",
+    "jade": "1.11.0",
+    "jade-loader": "0.7.1",
+    "joi": "6.6.1",
+    "jquery": "2.1.4",
+    "js-yaml": "3.4.1",
+    "json-loader": "0.5.3",
+    "json-stringify-safe": "5.0.1",
+    "jstimezonedetect": "1.0.5",
+    "leaflet": "0.7.5",
+    "less": "2.7.0",
+    "less-loader": "2.2.3",
+    "lodash": "3.10.1",
+    "marked": "0.3.6",
+    "minimatch": "2.0.10",
+    "mkdirp": "0.5.1",
+    "moment": "2.13.0",
+    "moment-timezone": "0.5.4",
+    "no-ui-slider": "1.2.0",
+    "node-fetch": "1.3.2",
+    "node-uuid": "1.4.7",
+    "pegjs": "0.9.0",
+    "querystring-browser": "1.0.4",
+    "raw-loader": "0.5.1",
+    "request": "2.61.0",
+    "rimraf": "2.4.3",
+    "rison-node": "1.0.0",
+    "rjs-repack-loader": "1.0.6",
+    "script-loader": "0.6.1",
+    "semver": "5.1.0",
+    "style-loader": "0.12.3",
+    "tar": "2.2.0",
+    "tinygradient": "0.3.0",
+    "trunc-html": "1.0.2",
+    "trunc-text": "1.0.2",
+    "url-loader": "0.5.6",
+    "validate-npm-package-name": "2.2.2",
+    "vision": "4.1.0",
+    "webpack": "1.12.15",
+    "webpack-directory-name-as-main": "1.0.0",
+    "whatwg-fetch": "0.9.0",
+    "wreck": "6.2.0"
   },
   "devDependencies": {
-    "bower": "^1.4.1",
-    "bower-license": "^0.2.6",
-    "event-stream": "^3.1.5",
-    "expect.js": "^0.3.1",
-    "grunt": "^0.4.5",
+    "@elastic/eslint-config-kibana": "0.0.3",
+    "angular-mocks": "1.4.7",
+    "auto-release-sinon": "1.0.3",
+    "babel-eslint": "4.1.8",
+    "chai": "3.5.0",
+    "cheerio": "0.22.0",
+    "chokidar": "1.6.0",
+    "chromedriver": "2.24.1",
+    "elasticdump": "2.1.1",
+    "eslint": "1.10.3",
+    "eslint-plugin-mocha": "1.1.0",
+    "event-stream": "3.3.2",
+    "expect.js": "0.3.1",
+    "faker": "1.1.0",
+    "grunt": "1.0.1",
+    "grunt-aws-s3": "0.14.5",
+    "grunt-babel": "5.0.1",
     "grunt-cli": "0.1.13",
-    "grunt-contrib-clean": "^0.6.0",
-    "grunt-contrib-compress": "^0.13.0",
-    "grunt-contrib-copy": "^0.8.0",
-    "grunt-contrib-jade": "^0.14.0",
-    "grunt-contrib-jshint": "^0.11",
-    "grunt-contrib-less": "^1.0.1",
-    "grunt-contrib-requirejs": "^0.4.4",
-    "grunt-contrib-watch": "^0.6.1",
-    "grunt-esvm": "^1.0.1",
-    "grunt-jscs": "^1.8.0",
-    "grunt-mocha": "^0.4.10",
-    "grunt-replace": "^0.7.9",
-    "grunt-run": "^0.3.0",
-    "grunt-s3": "^0.2.0-alpha.3",
-    "grunt-simple-mocha": "^0.4.0",
-    "html-entities": "^1.1.1",
-    "http-proxy": "^1.8.1",
-    "husky": "^0.8.1",
-    "istanbul": "^0.3.15",
-    "jade": "^1.8.2",
-    "license-checker": "^3.1.0",
-    "libesvm": "^1.0.1",
-    "load-grunt-config": "^0.7.0",
-    "marked": "^0.3.3",
-    "marked-text-renderer": "^0.1.0",
-    "mkdirp": "^0.5.0",
-    "mocha": "^2.2.5",
-    "nock": "^1.6.0",
-    "npm": "^2.11.0",
-    "opn": "^1.0.0",
-    "path-browserify": "0.0.0",
-    "portscanner": "^1.0.0",
-    "progress": "^1.1.8",
-    "requirejs": "^2.1.14",
-    "rjs-build-analysis": "0.0.3",
-    "simple-git": "^1.3.0",
-    "sinon": "^1.12.2",
-    "tar": "^2.1.1"
+    "grunt-contrib-clean": "1.0.0",
+    "grunt-contrib-copy": "0.8.1",
+    "grunt-esvm": "3.2.6",
+    "grunt-karma": "2.0.0",
+    "grunt-run": "0.6.0",
+    "grunt-simple-mocha": "0.4.0",
+    "gruntify-eslint": "1.0.1",
+    "gulp-sourcemaps": "1.7.3",
+    "handlebars": "4.0.5",
+    "husky": "0.8.1",
+    "image-diff": "1.6.0",
+    "intern": "3.2.3",
+    "istanbul-instrumenter-loader": "0.1.3",
+    "karma": "1.2.0",
+    "karma-chrome-launcher": "0.2.0",
+    "karma-coverage": "0.5.1",
+    "karma-firefox-launcher": "0.1.6",
+    "karma-ie-launcher": "0.2.0",
+    "karma-mocha": "0.2.0",
+    "karma-safari-launcher": "0.1.1",
+    "license-checker": "5.1.2",
+    "load-grunt-config": "0.19.2",
+    "makelogs": "3.0.2",
+    "marked-text-renderer": "0.1.0",
+    "mocha": "2.5.3",
+    "murmurhash3js": "3.0.1",
+    "ncp": "2.0.0",
+    "nock": "8.0.0",
+    "npm": "3.10.8",
+    "portscanner": "1.0.0",
+    "proxyquire": "1.7.10",
+    "simple-git": "1.37.0",
+    "sinon": "1.17.2",
+    "source-map": "0.5.6",
+    "source-map-support": "0.2.10",
+    "supertest": "1.2.0",
+    "supertest-as-promised": "2.0.2"
   },
   "engines": {
-    "node": "~0.10 || ~0.12",
-    "iojs": ">=1.5"
+    "node": "6.9.0",
+    "npm": "3.10.8"
   }
 }
