@@ -1,28 +1,97 @@
 module.exports = function (grunt) {
-  var join = require('path').join;
-  var rel = require('path').join.bind(null, grunt.config.get('root'));
-  var directory = join(grunt.config.get('root'), 'esvm');
-  var dataDir = join(directory, 'data_dir');
+  var resolve = require('path').resolve;
+  var directory = resolve(__dirname, '../../esvm');
+  var dataDir = resolve(directory, 'data_dir');
+  var serverConfig = require('../../test/server_config');
 
   return {
     options: {
-      directory: directory,
       branch: 'master',
-      fresh: true,
+      fresh: !grunt.option('esvm-no-fresh'),
       config: {
-        path: {
-          data: dataDir
-        },
-        network: {
-          host: '127.0.0.1'
-        },
-        marvel: {
-          agent: {
-            enabled: false
+        http: {
+          port: 9200
+        }
+      }
+    },
+    dev: {
+      options: {
+        directory: resolve(directory, 'dev'),
+        config: {
+          path: {
+            data: dataDir
+          },
+          cluster: {
+            name: 'esvm-dev'
           }
         }
       }
     },
-    dev: {}
+    test: {
+      options: {
+        directory: resolve(directory, 'test'),
+        purge: true,
+        config: {
+          http: {
+            port: serverConfig.servers.elasticsearch.port
+          },
+          cluster: {
+            name: 'esvm-test'
+          }
+        }
+      }
+    },
+    ui: {
+      options: {
+        directory: resolve(directory, 'test'),
+        purge: true,
+        config: {
+          http: {
+            port: serverConfig.servers.elasticsearch.port
+          },
+          cluster: {
+            name: 'esvm-ui'
+          }
+        }
+      }
+    },
+    withPlugins: {
+      options: {
+        version: '2.1.0',
+        directory: resolve(directory, 'withPlugins'),
+        plugins: [
+          'license',
+          'shield',
+          'marvel-agent',
+          'watcher'
+        ],
+        shield: {
+          users: [
+            {
+              username: 'kibana',
+              password: 'notsecure',
+              roles: ['kibana4_server']
+            },
+            {
+              username: 'user',
+              password: 'notsecure',
+              roles: ['kibana4', 'marvel']
+            },
+            {
+              username: 'admin',
+              password: 'notsecure',
+              roles: ['admin']
+            }
+          ]
+        },
+        config: {
+          marvel: {
+            agent: {
+              interval: '60s'
+            }
+          }
+        }
+      }
+    }
   };
 };
