@@ -2,40 +2,52 @@ module.exports = function (grunt) {
   return {
     options: {
       licenses: [
+        '(BSD-2-Clause OR MIT OR Apache-2.0)',
+        '(BSD-2-Clause OR MIT)',
+        '(MIT AND CC-BY-3.0)',
+        '(MIT OR Apache-2.0)',
+        'AFLv2.1',
+        'Apache',
+        'Apache 2.0',
+        'Apache License, v2.0',
+        'Apache*',
+        'Apache, Version 2.0',
+        'Apache-2.0',
+        'BSD',
+        'BSD New',
+        'BSD*',
+        'BSD-2-Clause',
+        'BSD-3-Clause',
+        'BSD-3-Clause AND MIT',
+        'BSD-3-Clause OR MIT',
+        'BSD-like',
+        'CC-BY',
+        'CC-BY-4.0',
+        'ISC',
         'MIT',
         'MIT*',
         'MIT/X11',
-        'new BSD, and MIT',
-        'BSD',
-        'BSD*',
-        'BSD New',
-        'BSD-like',
-        'BSD-2-Clause',
-        'BSD-3-Clause',
-        'Apache',
-        'Apache*',
-        'Apache v2',
-        'Apache 2.0',
-        'Apache2',
-        'Apache-2.0',
-        'Apache, Version 2.0',
-        'ISC',
+        'OFL-1.1 AND MIT',
+        'Public domain',
         'WTFPL',
-        'Public-Domain'
+        'new BSD, and MIT'
       ],
       overrides: {
         'assert-plus@0.1.5': ['MIT'],
-        'angular-bootstrap@0.10.0': ['MIT'],
-        'angular-ui-ace@0.2.3': ['MIT'],
+        'buffers@0.1.1': ['MIT/X11'],
+        'bytes@1.0.0': ['MIT'],
+        'color-name@1.0.0': ['UNLICENSE'],
         'commander@2.2.0': ['MIT'],
-        'leaflet@0.7.2': ['BSD-2-Clause'],
-        'moment-timezone@0.0.6': ['MIT'],
-        'zeroclipboard@2.2.0': ['MIT'],
-        'FileSaver@undefined': ['MIT'],
-        'cycle@1.0.3': ['Public-Domain'],
-        'pkginfo@0.2.3': ['MIT'],
+        'css-color-names@0.0.1': ['MIT'],
+        'css-parse@1.0.4': ['MIT'],
+        'css-stringify@1.0.5': ['MIT'],
+        'css@1.0.8': ['MIT'],
+        'delegate@3.0.1': ['MIT'],
+        'flatten@0.0.1': ['MIT'],
+        'indexof@0.0.1': ['MIT'],
+        'ripemd160@0.2.0': ['MIT'],
+        'select@1.0.6': ['MIT'],
         'uglify-js@2.2.5': ['BSD'],
-        'amdefine@0.1.1': ['BSD-3-Clause', 'MIT']
       }
     }
   };
