@@ -1,8 +1,20 @@
+require('../../test/mocha_setup');
+
 module.exports = {
   options: {
-    timeout: 2000,
+    timeout: 10000,
+    slow: 5000,
     ignoreLeaks: false,
-    reporter: 'dot'
+    reporter: 'dot',
+    globals: ['nil']
   },
-  all: { src: ['<%= root %>/test/unit/{server,tasks}/**/*.js'] }
+  all: {
+    src: [
+      'test/**/__tests__/**/*.js',
+      'src/**/__tests__/**/*.js',
+      'test/fixtures/__tests__/*.js',
+      '!src/**/public/**',
+      '!**/_*.js'
+    ]
+  }
 };
